// Sesion 6 y 7/curso-angular/src/app/shared/models/pasajero/pasajero.ts
export interface Pasajero {
    idPasajero: number;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    fechaNacimiento: string;
    idPais: number;
    telefono: string;
    email: string;
    clave: string;
    tipoDocumento: string;
    numDocumento: string;
}